replace the main entity list with a default group called "all"
i'm not sure why i originally added the main entity list (probably for UX reasons), but i'm removing it now to give people more control over the performance of the pool.

with a main entity list, you'll always have a list of entities that has to be added to/removed from, even if you're not using it. and now that there's multiple sorting modes, i don't think i can choose a default sorting mode for that list that would be the right mode for everyone's situation. so by removing the main entity list and making it a default group instead, i'm giving people the option to customize the "main" group or opt out of it entirely.

some side effects of this change:
- previous "add" and "remove" events are removed
- "addToGroup" and "removeFromGroup" events are renamed to "add" and "remove"
- nata.oop requires a group name as the first argument